remove a circular dependency between approvalGate and Goal
