+++ src/common/goals/Goal.ts
@@ -3,13 +3,12 @@ import {
     ProjectOperationCredentials,
     TokenCredentials,
 } from "@atomist/automation-client/operations/common/ProjectOperationCredentials";
-import { StatusState } from "../../typings/types";
 import { createStatus } from "../../util/github/ghub";
 
 import { logger } from "@atomist/automation-client";
 import { RemoteRepoRef } from "@atomist/automation-client/operations/common/RepoId";
 import { requiresApproval } from "../../handlers/events/delivery/verify/approvalGate";
-import { BaseContext, GitHubStatusContext, GoalEnvironment } from "./gitHubContext";
+import { BaseContext, GitHubStatusAndFriends, GitHubStatusContext, GoalEnvironment } from "./gitHubContext";
 
 export interface GoalDefinition {
     environment: GoalEnvironment;
@@ -152,17 +151,6 @@ function setPendingStatus(id: GitHubRepoRef, context: GitHubStatusContext,
         context,
         description,
     });
-}
-
-export interface GitHubStatus {
-    context?: GitHubStatusContext;
-    description?: string;
-    state?: StatusState;
-    targetUrl?: string;
-}
-
-export interface GitHubStatusAndFriends extends GitHubStatus {
-    siblings: GitHubStatus[];
 }
 
 export function currentGoalIsStillPending(currentGoal: GitHubStatusContext, status: GitHubStatusAndFriends): boolean {
