+++ src/util/github/ghub.ts
@@ -1,12 +1,9 @@
 import { logger } from "@atomist/automation-client";
 import { GitHubRepoRef, isGitHubRepoRef } from "@atomist/automation-client/operations/common/GitHubRepoRef";
-import { RemoteRepoRef, RepoRef } from "@atomist/automation-client/operations/common/RepoId";
+import { RemoteRepoRef } from "@atomist/automation-client/operations/common/RepoId";
 import { Issue } from "@atomist/automation-client/util/gitHub";
 import { doWithRetry } from "@atomist/automation-client/util/retry";
-import axios, {
-    AxiosPromise,
-    AxiosRequestConfig,
-} from "axios";
+import axios, { AxiosPromise, AxiosRequestConfig, } from "axios";
 
 export type State = "error" | "failure" | "pending" | "success";
